Remove file nesting
Remove the layer of nesting for all component and page files. No longer since none of them has a companion CSS file anymore.
Adjust import paths to suit.
