+++ src/components/GetUsersButton.jsx
@@ -1,17 +1,20 @@
 import firebase from "firebase/app";
 
-import { API_URL } from "./../../util/urls";
+import { useAuth } from "./../util/useAuth";
+import { API_URL } from "./../util/urls";
 
 
-function SetRoleButton () {
+function GetUsersButton () {
+	const auth = useAuth();
 
-	function setRole () {
+	function getUsers () {
+		console.log(auth);
 		firebase.auth().currentUser.getIdToken(true).then(function (token) {
 			const requestOptions = {
-				method: "POST",
+				method: "GET",
 				headers: { authorization: `Bearer ${token}` }
 			};
-			fetch(`${API_URL}/loginsuccess/setrole`, requestOptions)
+			fetch(`${API_URL}/users`, requestOptions)
 				.then(res => res.json())
 				.then(data => console.log(data));
 		}).catch(function (error) {
@@ -20,10 +23,10 @@ function SetRoleButton () {
 	}
 
 	return (
-		<button onClick={setRole}>
-			Set Role
+		<button onClick={getUsers}>
+			Get Users
 		</button>
 	);
 }
 
-export default SetRoleButton;
+export default GetUsersButton;
